requirements.txt
django==4.2.7
graphene_django==3.1.5
graphene==3.3
google-auth==2.23.4
google-cloud-bigquery==3.13.0
pandas==2.1.3
plotly==5.18.0